Add CMakeLists.txt
This commit adds CMakeLists.txt, allowing users to build libcsp
against FreeRTOS with CMake. Unlike Meson, CMake doesn't enforce
placing libcsp under subprojects/. However, this CMakeLists.txt
assumes that you have it there. This is done to align with the setup
in meson.build, not due to a limitation in CMake.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>