cmake: Leverage FreeRTOS's CMake Support
FreeRTOS-Kernel added CMake support in cdd406a62f31f. This commit
utilizes the CMake support for the demo app.

Note that CMAKE_CURRENT_SOURCE_DIR is specified for freertos_config as
SYSTEM because I want to ensure that all source and header files find
FreeRTOSConfig.h.

It still targets the POSIX port. If you intend to build it for another
PORT, you will need a toolchain file for your target (and probably a libc
to go along with it).

Ref:
 - https://github.com/FreeRTOS/FreeRTOS-Kernel/pull/421
 - https://cmake.org/cmake/help/latest/manual/cmake-toolchains.7.html

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>